5 coding host side of actor sys

+++ src/actors/shared_emu.h
@@ -24,7 +24,7 @@ typedef uint16_t bj_core_id_t;	// e_coreid_t
 typedef uint16_t bj_core_co_t;
 typedef uint16_t bj_core_nn_t;
 
-typedef uint16_t bj_size_t;
+typedef uint32_t bj_size_t;
 
 #define bj_addr_val_in_p16(p16) ((bj_addr_t)(bj_v32_of_p16(p16)))
